refactor: migrate math utils into bp_iters
